requirements.txt
numpy
scipy
sklearn
h5py>=3.1.0
ruamel.yaml
attrs
gstools
memoization
tensorflow
pandas
